refactor(content): move GetContent into its own file
