+++ internal/github/repo_test.go
@@ -1,8 +1,6 @@
 package github
 
 import (
-	"encoding/base64"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
@@ -37,52 +35,4 @@ func TestGetDefaultBranch(t *testing.T) {
 	if repo.DefaultBranch != "" {
 		t.Fatalf("expected repo not to change got %v", repo)
 	}
-}
-
-func TestGetContent(t *testing.T) {
-	t.Parallel()
-
-	repo := Repo{Owner: User{"owner"}, Repo: "repo"}
-
-	t.Run("fails to decode the content", func(t *testing.T) {
-		t.Parallel()
-
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path != "/repos/owner/repo/contents/path/to/file" {
-				t.Fatal("invalid path", r.URL.Path)
-			}
-
-			fmt.Fprintln(w, `{"content": "_not base64"}`)
-		}))
-
-		g := New("token", ts.URL)
-
-		_, err := g.GetContent(t.Context(), repo, "path/to/file")
-		if !errors.Is(err, base64.CorruptInputError(0)) {
-			t.Fatalf("expected base64 error, got %v", err)
-		}
-	})
-
-	t.Run("succeeds", func(t *testing.T) {
-		t.Parallel()
-
-		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path != "/repos/owner/repo/contents/path/to/file" {
-				t.Fatal("invalid path", r.URL.Path)
-			}
-
-			fmt.Fprintln(w, `{"content": "b2s="}`)
-		}))
-
-		g := New("token", ts.URL)
-
-		c, err := g.GetContent(t.Context(), repo, "path/to/file")
-		if err != nil {
-			t.Fatalf("expected no error, got %v", err)
-		}
-
-		if c.Content != "ok" {
-			t.Fatalf("expected content to be 'ok' but got %s", c.Content)
-		}
-	})
 }
